addressed PR notes
shared functions in separate module
simplified and moving more towards shell script style python script
using python3 constucts for better code
